Modifies: Styling to enhance UI experience   1. Warning colors added to buttons that can delete previous 	input.   2. Modified color scheme to make the design more fluid.
Corrects: Float operations.
  1. User can now add floats and not suffer a loss in precision.

+++ js/js.js
@@ -43,11 +43,11 @@ $(".keypad").on("click", function() {
         a *= -1;
         $("#screen").text(a);
     } else if ($(this).hasClass("col keypad equals")) {
-        a = parseInt(a);
+        a = parseFloat(a);
         checkOperation(result, oper, a);
     } else {
         // parse the first input
-        result = parseInt(a)
+        result = parseFloat(a)
         oper = this.innerText;
         //handle special cases for clear, del, and clear all
         if (oper === "CE") {
